+++ controllers/courses.js
@@ -2,6 +2,7 @@ const Course = require("../models/Course");
 const Bootcamp = require("../models/Bootcamp");
 const ErrorResponse = require("../utils/errorResponse");
 const asyncHandler = require("../middlewares/async");
+const { isTheOwner } = require("../utils/user");
 
 // @desc Get all courses
 // @route GET /api/v1/courses
@@ -44,13 +45,25 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @access Private
 exports.addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
 
+  // Check for bootcamp existence
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
   if (!bootcamp) {
     return next(
       new ErrorResponse(
         `No bootcamp with the id of ${req.params.bootcampId}`,
         404
+      )
+    );
+  }
+
+  // Make sure user is bootcamp owner
+  if (!isTheOwner(req.user, bootcamp.user)) {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+        401
       )
     );
   }
@@ -71,6 +84,16 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
     return next(new ErrorResponse(`No course with the id of ${id}`, 404));
   }
 
+  // Make sure user is course owner
+  if (!isTheOwner(req.user, course.user)) {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update course ${course._id}`,
+        401
+      )
+    );
+  }
+
   // Update the course
   const updatedCourse = await Course.findByIdAndUpdate(id, req.body, {
     new: true,
@@ -89,6 +112,16 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
 
   if (!course) {
     return next(new ErrorResponse(`No course with the id of ${id}`, 404));
+  }
+
+  // Make sure user is course owner
+  if (!isTheOwner(req.user, course.user)) {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete course ${course._id}`,
+        401
+      )
+    );
   }
 
   await course.deleteOne();
